+++ airflow/executors/base_executor.py
@@ -1,7 +1,7 @@
 from builtins import range
 
 from airflow import configuration
-from airflow.utils.classes import State
+from airflow.utils.state import State
 from airflow.utils.logging import LoggingMixin
 
 PARALLELISM = configuration.getint('core', 'PARALLELISM')
