+++ airflow/executors/local_executor.py
@@ -6,7 +6,7 @@ from builtins import range
 
 from airflow import configuration
 from airflow.executors.base_executor import BaseExecutor
-from airflow.utils.classes import State
+from airflow.utils.state import State
 from airflow.utils.logging import LoggingMixin
 
 PARALLELISM = configuration.get('core', 'PARALLELISM')
