+++ airflow/operators/python_operator.py
@@ -3,7 +3,7 @@ from datetime import datetime
 import logging
 
 from airflow.models import BaseOperator, TaskInstance
-from airflow.utils.classes import State
+from airflow.utils.state import State
 from airflow.utils.decorators import apply_defaults
 from airflow import settings
