+++ airflow/operators/sensors.py
@@ -11,7 +11,7 @@ from airflow import hooks, settings
 from airflow.exceptions import AirflowException, AirflowSensorTimeout
 from airflow.models import BaseOperator, TaskInstance, Connection as DB
 from airflow.hooks import BaseHook
-from airflow.utils.classes import State
+from airflow.utils.state import State
 from airflow.utils.decorators import apply_defaults
