--- airflow/utils/classes.py
@@ -1,107 +1,0 @@
-from __future__ import absolute_import
-from __future__ import division
-from __future__ import print_function
-from __future__ import unicode_literals
-
-from builtins import object
-
-import logging
-import signal
-
-
-class AirflowTaskTimeout(Exception):
-    pass
-
-
-class timeout(object):
-    """
-    To be used in a ``with`` block and timeout its content.
-    """
-    def __init__(self, seconds=1, error_message='Timeout'):
-        self.seconds = seconds
-        self.error_message = error_message
-
-    def handle_timeout(self, signum, frame):
-        logging.error("Process timed out")
-        raise AirflowTaskTimeout(self.error_message)
-
-    def __enter__(self):
-        try:
-            signal.signal(signal.SIGALRM, self.handle_timeout)
-            signal.alarm(self.seconds)
-        except ValueError as e:
-            logging.warning("timeout can't be used in the current context")
-            logging.exception(e)
-
-    def __exit__(self, type, value, traceback):
-        try:
-            signal.alarm(0)
-        except ValueError as e:
-            logging.warning("timeout can't be used in the current context")
-            logging.exception(e)
-
-
-class TriggerRule(object):
-    ALL_SUCCESS = 'all_success'
-    ALL_FAILED = 'all_failed'
-    ALL_DONE = 'all_done'
-    ONE_SUCCESS = 'one_success'
-    ONE_FAILED = 'one_failed'
-    DUMMY = 'dummy'
-
-    @classmethod
-    def is_valid(cls, trigger_rule):
-        return trigger_rule in cls.all_triggers()
-
-    @classmethod
-    def all_triggers(cls):
-        return [getattr(cls, attr)
-                for attr in dir(cls)
-                if not attr.startswith("__") and not callable(getattr(cls, attr))]
-
-
-class State(object):
-    """
-    Static class with task instance states constants and color method to
-    avoid hardcoding.
-    """
-    QUEUED = "queued"
-    RUNNING = "running"
-    SUCCESS = "success"
-    SHUTDOWN = "shutdown"  # External request to shut down
-    FAILED = "failed"
-    UP_FOR_RETRY = "up_for_retry"
-    UPSTREAM_FAILED = "upstream_failed"
-    SKIPPED = "skipped"
-
-    state_color = {
-        QUEUED: 'gray',
-        RUNNING: 'lime',
-        SUCCESS: 'green',
-        SHUTDOWN: 'blue',
-        FAILED: 'red',
-        UP_FOR_RETRY: 'gold',
-        UPSTREAM_FAILED: 'orange',
-        SKIPPED: 'pink',
-    }
-
-    @classmethod
-    def color(cls, state):
-        if state in cls.state_color:
-            return cls.state_color[state]
-        else:
-            return 'white'
-
-    @classmethod
-    def color_fg(cls, state):
-        color = cls.color(state)
-        if color in ['green', 'red']:
-            return 'white'
-        else:
-            return 'black'
-
-    @classmethod
-    def runnable(cls):
-        return [
-            None, cls.FAILED, cls.UP_FOR_RETRY, cls.UPSTREAM_FAILED,
-            cls.SKIPPED, cls.QUEUED]
