+++ airflow/utils/state.py
@@ -1,0 +1,50 @@
+from __future__ import unicode_literals
+
+from builtins import object
+
+
+class State(object):
+    """
+    Static class with task instance states constants and color method to
+    avoid hardcoding.
+    """
+    QUEUED = "queued"
+    RUNNING = "running"
+    SUCCESS = "success"
+    SHUTDOWN = "shutdown"  # External request to shut down
+    FAILED = "failed"
+    UP_FOR_RETRY = "up_for_retry"
+    UPSTREAM_FAILED = "upstream_failed"
+    SKIPPED = "skipped"
+
+    state_color = {
+        QUEUED: 'gray',
+        RUNNING: 'lime',
+        SUCCESS: 'green',
+        SHUTDOWN: 'blue',
+        FAILED: 'red',
+        UP_FOR_RETRY: 'gold',
+        UPSTREAM_FAILED: 'orange',
+        SKIPPED: 'pink',
+    }
+
+    @classmethod
+    def color(cls, state):
+        if state in cls.state_color:
+            return cls.state_color[state]
+        else:
+            return 'white'
+
+    @classmethod
+    def color_fg(cls, state):
+        color = cls.color(state)
+        if color in ['green', 'red']:
+            return 'white'
+        else:
+            return 'black'
+
+    @classmethod
+    def runnable(cls):
+        return [
+            None, cls.FAILED, cls.UP_FOR_RETRY, cls.UPSTREAM_FAILED,
+            cls.SKIPPED, cls.QUEUED]
